Rearrange callbacks code into typical Django modules structure
